requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
amadeus==6.0.1
reportlab==4.0.4
Flask-CORS==4.0.0
